+++ drizzle/0006_awesome_doctor_doom.sql
@@ -1,0 +1,2 @@
+ALTER TABLE `providers` ADD `last_image_sha256` text;--> statement-breakpoint
+ALTER TABLE `providers` ADD `last_image_parsed` text;